// betterbikes-server/src/interfaces/vehiclePost.ts
export interface IVehiclePost{ 
    vehicleImage?: string;
    vehicleName: string;
    vehicleBrand: string;
    vehicleType: string;
    vehicleColor: string;
    vehiclePrice: string;
    vehicleNumber: string;
    vehicleAddress: string;
    vehicleManufactureDate: string;
    vehicleDescription: string;
    vehicleFeatures: string[]
}